Struct coruscant_world::level_dat::Data

pub struct Data {
    pub custom_boss_enevts: HashMap<String, BossEvent>,
    pub data_packs: DataPacks,
    pub dimension_data: DimensionData,
    pub version_int: i32,
    pub initialized: bool,
    pub level_name: String,
    pub generator_name: String,
    pub random_seed: i64,
    pub map_features: bool,
    pub last_played: i64,
    pub size_on_disk: i64,
    pub allow_commands: bool,
    pub hardcore: bool,
    pub data_version: i32,
    pub difficulty: i8,
    pub difficulty_locked: bool,
    pub time: i64,
    pub day_time: i64,
    pub game_type: i32,
    pub spawn_x: i32,
    pub spawn_y: i32,
    pub spawn_z: i32,
    pub border_center_x: f64,
    pub border_center_z: f64,
    pub border_size: f64,
    pub border_safe_zone: f64,
    pub border_warning_blocks: f64,
    pub border_warning_time: f64,
    pub border_size_lerp_target: f64,
    pub border_size_lerp_time: i64,
    pub border_damage_per_block: f64,
    pub raining: bool,
    pub rain_time: i32,
    pub thundering: bool,
    pub thunder_time: i32,
    pub clear_weather_time: i32,
    pub wandering_trader_id: String,
    pub wandering_trader_spawn_chance: i32,
    pub wanding_trader_spawn_delay: i32,
    pub version: Version,
}

This tag contains all the level data.

Fields
custom_boss_enevts: HashMap<String, BossEvent>

A ID/BossEvent collection of bossbars. ID of a bossbar is a string like custom::boss.

data_packs: DataPacks

Options for datapacks.

dimension_data: DimensionData

This tag contains level data specific to certain dimensions.

version_int: i32

The NBT version of the level, 19133.

initialized: bool

Normally true after a world has been initialized properly after creation. If the initial simulation was canceled somehow, this can be false and the world will be re-initialized on next load.

level_name: String

The name of the level.

generator_name: String

The name of the generator; "default", "flat", "largeBiomes", "amplified", "buffet", or "debug_all_block_states".

Not case sensitive, but always written in the case here.

random_seed: i64

The random level seed used to generate consistent terrain.

map_features: bool

True if the map generator should place structures such as villages, strongholds, and mineshafts. Defaults to true. Always true if the world type is Customized.

last_played: i64

The Unix time in milliseconds when the level was last loaded.

size_on_disk: i64

The estimated size in bytes of the level. Currently not modified or used by Minecraft, but was previously.

allow_commands: bool

True if cheats are enabled.

hardcore: bool

True if the player will respawn in Spectator on death in singleplayer. Affects all three game modes.

data_version: i32

An integer displaying the data version.

difficulty: i8

The current difficulty setting. 0 is Peaceful, 1 is Easy, 2 is Normal, and 3 is Hard. Defaults to 2.

difficulty_locked: bool

True if the difficulty has been locked. Defaults to false.

time: i64

The number of ticks since the start of the level.

day_time: i64

The time of day. 0 is sunrise, 6000 is mid day, 12000 is sunset, 18000 is mid night, 24000 is the next day's 0. This value keeps counting past 24000 and does not reset to 0.

game_type: i32

The default game mode for the singleplayer player when they initially spawn. 0 is Survival, 1 is Creative, 2 is Adventure, 3 is Spectator.

Note: Singleplayer worlds do not use this field to save which game mode the player is currently in.

spawn_x: i32

The X coordinate of the world spawn.

spawn_y: i32

The Y coordinate of the world spawn.

spawn_z: i32

The Z coordinate of the world spawn.

border_center_x: f64

Center of the world border on the X coordinate. Defaults to 0.

border_center_z: f64

Center of the world border on the Z coordinate. Defaults to 0.

border_size: f64

Width of the border. Defaults to 60000000.

border_safe_zone: f64

Defaults to 5. // todo: figure out what these are

border_warning_blocks: f64

Defaults to 5.

border_warning_time: f64

Defaults to 15.

border_size_lerp_target: f64

Defaults to 60000000.

border_size_lerp_time: i64

Defaults to 0.

border_damage_per_block: f64

Defaults to 0.2.

raining: bool

True if the level is currently experiencing rain, snow, and cloud cover.

rain_time: i32

The number of ticks before "raining" is toggled and this value gets set to another random value.

thundering: bool

True if the rain/snow/cloud cover is a lightning storm and dark enough for mobs to spawn under the sky.

thunder_time: i32

The number of ticks before "thundering" is toggled and this value gets set to another random value.

clear_weather_time: i32

The number of ticks until "clear weather" has ended.

wandering_trader_id: String

The UUID of the current wandering trader in the world.

wandering_trader_spawn_chance: i32

The current chance of the wandering trader spawning next attempt.

This value is the percentage and will be divided by 10 when loaded by the game, for example a value of 50 means 5.0% chance.

wanding_trader_spawn_delay: i32

The amount of ticks until another wandering trader is attempted to spawn.

version: Version

Information about the Minecraft version the world was saved in.